mv : get_head_token_str() from builtin to tokenize
